+++ Common/Code/Node.cpp
@@ -24,7 +24,7 @@ std::string Node::friendlyDescription() {
             //if the description starts with "signet-as signet..." or "signet signet..."
             //then strip the redundant first word.
             bool stripFirst = false;
-            int firstWordEnd = rawTextDescription.find(' ');
+            long firstWordEnd = rawTextDescription.find(' ');
             if (firstWordEnd != std::string::npos) {
                 std::string firstWord = rawTextDescription.substr(0, firstWordEnd);
                 if (firstWord.find('-') != std::string::npos) {
@@ -32,9 +32,9 @@ std::string Node::friendlyDescription() {
                     stripFirst = true;
                 } else {
                     //compare to second word
-                    int secondWordStart = firstWordEnd + 1; //note: assuming singlespacing
-                    int secondWordEnd = rawTextDescription.find(' ', secondWordStart);
-                    int secondWordLen = (secondWordEnd == std::string::npos) ? std::string::npos : (secondWordEnd-secondWordStart);
+                    long secondWordStart = firstWordEnd + 1; //note: assuming singlespacing
+                    long secondWordEnd = rawTextDescription.find(' ', secondWordStart);
+                    long secondWordLen = (secondWordEnd == std::string::npos) ? std::string::npos : (secondWordEnd-secondWordStart);
                     if (rawTextDescription.compare(secondWordStart, secondWordLen, firstWord) == 0) {
                         stripFirst = true;
                     }
@@ -50,9 +50,9 @@ std::string Node::friendlyDescription() {
             //LOG("before upper: %s", mFriendlyDescription.c_str());
             //now capitalize every word.
             //note: assuming singlespacing (and ascii)
-            int prevWordEnd = -1;
+            long prevWordEnd = -1;
             do {
-                int wordStart = prevWordEnd + 1;
+                long wordStart = prevWordEnd + 1;
                 mFriendlyDescription[wordStart] = toupper(mFriendlyDescription[wordStart]);
                 prevWordEnd = mFriendlyDescription.find(' ', wordStart);
             } while (prevWordEnd != std::string::npos);
